+++ src/Zybo-Z7-20-HDMI/src/video_capture/video_capture.h
@@ -138,8 +138,8 @@ typedef struct {
 		u32 stride; /* The line stride of the framebuffers, in bytes */
 		u32 curFrame; /* Current frame being displayed */
 		XGpio gpio; /* XGPIO driver struct */
-		u16 vtcId; /* Device ID of VTC core as defined in xparameters.h */
-		u16 vtcIrptId; /* Interrupt ID for the VTC core */
+		u32 vtcId; /* Device ID of VTC core as defined in xparameters.h */
+		u32 vtcIrptId; /* Interrupt ID for the VTC core */
 		u32 startOnDetect; /* boolean Flag indicating whether or not the VDMA should be started in the interrupt when a signal is detected */
 		VideoState state; /* Indicates if the Display is currently running */
 } VideoCapture;
@@ -154,7 +154,7 @@ typedef struct {
 
 int VideoStop(VideoCapture *videoPtr);
 int VideoStart(VideoCapture *videoPtr);
-int VideoInitialize(VideoCapture *videoPtr, INTC *intCtrl, XAxiVdma *vdma, u16 gpioId, u16 vtcId, u32 vtcIrptId, u8 *framePtr[VIDEO_NUM_FRAMES], u32 stride, u32 startOnDet);
+int VideoInitialize(VideoCapture *videoPtr, INTC *intCtrl, XAxiVdma *vdma, u32 gpioId, u32 vtcId, u32 vtcIrptId, u8 *framePtr[VIDEO_NUM_FRAMES], u32 stride, u32 startOnDet);
 int VideoChangeFrame(VideoCapture *videoPtr, u32 frameIndex);
 void VideoSetCallback(VideoCapture *videoPtr, VideoCallBack CallBackFunc, void *CallBackRef);
 void GpioIsr(void *InstancePtr);
